+++ tests/conftest.py
@@ -21,6 +21,7 @@ from src.custom_types import (
     CompressEmbeddingsConfig,
     EncodingTextConfig,
     DateExtractionConfig,
+    NumericalFeaturesConfig,
 )
 from src.bigquery_connector.bigquery_connector import BigQueryConnector
 from src.data_preparation.data_preparation import StackOverflowDataPreparation
@@ -214,7 +215,7 @@ def fixture_sentence_transformers_config(
 ) -> SentenceTransformersConfig:
     """
     Fixture for a SentenceTransformersConfig object
-    from src/types.SentenceTransformersConfig class definition.
+    from src/custom_types.SentenceTransformersConfig class definition.
 
     Args:
         sentence_transformers_config (Dictionary): Configurations for a SentenceTransformersConfig object
@@ -232,7 +233,7 @@ def fixture_embeddings_config(
 ) -> EmbeddingsConfig:
     """
     Fixture for an EmbeddingsConfig object
-    from src/types.EmbeddingsConfig class definition.
+    from src/custom_types.EmbeddingsConfig class definition.
 
     Args:
         fixture_sentence_transformers_config (SentenceTransformersConfig): Configurations for a SentenceTransformersConfig object
@@ -251,7 +252,7 @@ def fixture_embeddings_config(
 def fixture_pca_config(pca_config: dict = config["data_preparation"]["pca_config"]) -> PCAConfig:
     """
     Fixture for a PCAConfig object
-    from src/types.PCAConfig class definition.
+    from src/custom_types.PCAConfig class definition.
 
     Args:
         pca_config (Dictionary): PCA configurations
@@ -269,7 +270,7 @@ def fixture_compress_embeddings_config(
 ) -> CompressEmbeddingsConfig:
     """
     Fixture for a CompressEmbeddingsConfig object
-    from src/types.CompressEmbeddingsConfig class definition.
+    from src/custom_types.CompressEmbeddingsConfig class definition.
 
     Args:
         fixture_pca_config (PCAConfig): Configurations for a PCAConfig object
@@ -290,7 +291,7 @@ def fixture_encode_text_config(
 ) -> EncodingTextConfig:
     """
     Fixture for an EncodeTextConfig object
-    from src/types.EncodingTextConfig class definition.
+    from src/custom_types.EncodingTextConfig class definition.
 
     Args:
         fixture_embeddings_config (EmbeddingsConfig): Configuration for embedding generation
@@ -338,7 +339,7 @@ def fixture_date_extraction_config(
 ) -> DateExtractionConfig:
     """
     Fixture for a DateExtractionConfig object
-    from src/types.DateExtractionConfig class definition.
+    from src/custom_types.DateExtractionConfig class definition.
 
     Args:
         date_extraction_config (Dictionary): Configurations for a DateExtractionConfig object
@@ -347,3 +348,20 @@ def fixture_date_extraction_config(
         (DateExtractionConfig): Object with all configs for a DateExtractionConfig set to True
     """
     return DateExtractionConfig(**date_extraction_config)
+
+
+@pytest.fixture
+def fixture_numerical_features_config(
+    numerical_features_config: dict = config["data_preparation"]["numerical_features_config"],
+) -> NumericalFeaturesConfig:
+    """
+    Fixture for a NumericalFeaturesConfig object
+    from src/custom_types.NumericalFeaturesConfig class definition.
+
+    Args:
+        numerical_features_config (Dictionary): Numerical feature transformation configurations
+
+    Returns:
+        (NumericalFeaturesConfig): Object including numerical feature transformation configurations
+    """
+    return NumericalFeaturesConfig(**numerical_features_config)